% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.explorer-package.R
\docType{package}
\name{cluster.explorer-package}
\alias{cluster.explorer}
\alias{cluster.explorer-package}
\title{cluster.explorer: Interactive Cluster Explorer for Neuroimaging Data}
\description{
Provides an interactive Shiny application for exploring statistical clusters in neuroimaging data. Supports volumetric and surface-based visualisation, atlas-annotated cluster tables, design variable plots, and extensible analysis plugins. The computation engine has zero Shiny dependencies and can be used programmatically.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbuchsbaum/neuroatlas}
  \item Report bugs at \url{https://github.com/bbuchsbaum/neuroatlas/issues}
}

}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{brad.buchsbaum@gmail.com} (\href{https://orcid.org/0000-0002-1108-4866}{ORCID})

}
\keyword{internal}
