% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_class.R
\name{export_csv}
\alias{export_csv}
\alias{export_csv.cluster_report_result}
\title{Export Cluster Report Results to CSV}
\usage{
export_csv(x, ...)

\method{export_csv}{cluster_report_result}(x, dir = ".", prefix = "cluster_report", ...)
}
\arguments{
\item{x}{A `cluster_report_result` object.}

\item{...}{Ignored.}

\item{dir}{Directory to write into. Default `"."`.}

\item{prefix}{Filename prefix. Default `"cluster_report"`.}
}
\value{
Invisibly, a character vector of written file paths.
}
\description{
Write cluster tables and time-course data as CSV files suitable for
supplementary materials.
}
