% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_class.R
\name{print.cluster_report_result}
\alias{print.cluster_report_result}
\title{Print a Cluster Report Result}
\usage{
\method{print}{cluster_report_result}(x, ...)
}
\arguments{
\item{x}{A `cluster_report_result` object.}

\item{...}{Ignored.}
}
\description{
Print a Cluster Report Result
}
