% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_class.R
\name{summary.cluster_report_result}
\alias{summary.cluster_report_result}
\title{Summarize a Cluster Report Result}
\usage{
\method{summary}{cluster_report_result}(object, ...)
}
\arguments{
\item{object}{A `cluster_report_result` object.}

\item{...}{Ignored.}
}
\description{
Summarize a Cluster Report Result
}
